requirements.txt
# Pacotes necessários para análise de simulações
cassandra-driver>=3.25.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
scipy>=1.10.0
scikit-learn>=1.3.0
lxml>=4.9.0
jupyter>=1.0.0
notebook>=6.5.0